-tipo de limite de dias de hecho pasa de int a long -agrego fechasAcontecimientos coherentes en los tests.

+++ src/main/java/ar/edu/utn/frba/dds/metamapa/models/entities/Hecho.java
@@ -18,7 +18,7 @@ import java.util.List;
 @Setter
 public class Hecho {
 
-  private int limiteDiasEdicion;
+  private long limiteDiasEdicion;
 
   private Long id;
